use macro for realm functions
