+++ apps/sim/lib/sim/realm.ex
@@ -2,63 +2,51 @@ defmodule Sim.Realm do
   @moduledoc """
   The context for realm
   """
-
   alias Sim.Realm.Server
 
-  def get_root() do
-    GenServer.call(Server, :get_root)
-  end
+  defmacro __using__(opts) do
+    app_name = opts[:name]
 
-  def set_root(root) do
-    GenServer.call(Server, {:set_root, root})
-  end
+    quote do
+      @supervisor unquote(app_name)
+      @server Server.server_name(unquote(app_name))
 
-  def create(factory, config) when is_atom(factory) do
-    GenServer.call(Server, {:create, factory, config})
-  end
+      def get_root() do
+        GenServer.call(@server, :get_root)
+      end
 
-  def recreate(factory, config) when is_atom(factory) do
-    GenServer.call(Server, :stop_sim)
-    GenServer.call(Server, {:create, factory, config})
-  end
+      def set_root(root) do
+        GenServer.call(@server, {:set_root, root})
+      end
 
-  def restart() do
-    :ok = Sim.Realm.Supervisor.restart_realm()
-  end
+      def create(factory, config) when is_atom(factory) do
+        GenServer.call(@server, {:create, factory, config})
+      end
 
-  def start_sim(func) do
-    GenServer.call(Server, {:start_sim, func})
-  end
+      def recreate(factory, config) when is_atom(factory) do
+        GenServer.call(@server, :stop_sim)
+        GenServer.call(@server, {:create, factory, config})
+      end
 
-  def stop_sim() do
-    GenServer.call(Server, :stop_sim)
-  end
+      def restart() do
+        :ok = @supervisor.restart_realm()
+      end
 
-  def started?() do
-    GenServer.call(Server, :started?)
-  end
+      def start_sim(func) do
+        GenServer.call(@server, {:start_sim, func})
+      end
 
-  def sim(func) do
-    GenServer.cast(Server, {:sim, func})
-  end
+      def stop_sim() do
+        GenServer.call(@server, :stop_sim)
+      end
 
-  # game of life specific
+      def started?() do
+        GenServer.call(@server, :started?)
+      end
 
-  def start_sim() do
-    start_sim(fn data -> GameOfLife.Simulation.sim(data) end)
-  end
-
-  def create(size) do
-    create(Grid, %{size: size})
-  end
-
-  def recreate(size) do
-    recreate(Grid, %{size: size})
-  end
-
-  def clear() do
-  end
-
-  def toggle() do
+      def sim(func) do
+        GenServer.cast(@server, {:sim, func})
+      end
+    end
   end
 end
